[199] ComboBox changes
